# tsdb chunk reading: faster CRC check by avoiding allocations