Instead of creating a new hashing object every time, call `crc32.Checksum` which computes the answer without allocations.

I noticed this in compactor benchmarks (incorporating #10788 otherwise the benchmark won't run on my machine): 
```
name                                                                            old time/op    new time/op    delta
Compaction/type=normal,blocks=4,series=10000,samplesPerSeriesPerBlock=101-4        471ms ± 0%     468ms ± 0%   -0.51%  (p=1.000 n=1+1)
Compaction/type=normal,blocks=4,series=10000,samplesPerSeriesPerBlock=1001-4       1.26s ± 0%     1.20s ± 0%   -5.04%  (p=1.000 n=1+1)
Compaction/type=normal,blocks=4,series=10000,samplesPerSeriesPerBlock=2001-4       1.97s ± 0%     1.71s ± 0%  -12.75%  (p=1.000 n=1+1)
Compaction/type=normal,blocks=4,series=10000,samplesPerSeriesPerBlock=5001-4       3.93s ± 0%     3.61s ± 0%   -8.12%  (p=1.000 n=1+1)
Compaction/type=vertical,blocks=4,series=10000,samplesPerSeriesPerBlock=101-4      1.64s ± 0%     1.64s ± 0%   +0.24%  (p=1.000 n=1+1)
Compaction/type=vertical,blocks=4,series=10000,samplesPerSeriesPerBlock=1001-4     8.64s ± 0%     8.49s ± 0%   -1.80%  (p=1.000 n=1+1)
Compaction/type=vertical,blocks=4,series=10000,samplesPerSeriesPerBlock=2001-4     15.8s ± 0%     15.4s ± 0%   -2.52%  (p=1.000 n=1+1)
Compaction/type=vertical,blocks=4,series=10000,samplesPerSeriesPerBlock=5001-4     37.5s ± 0%     36.4s ± 0%   -3.08%  (p=1.000 n=1+1)

name                                                                            old alloc/op   new alloc/op   delta
Compaction/type=normal,blocks=4,series=10000,samplesPerSeriesPerBlock=101-4       86.2MB ± 0%    85.2MB ± 0%   -1.15%  (p=1.000 n=1+1)
Compaction/type=normal,blocks=4,series=10000,samplesPerSeriesPerBlock=1001-4       125MB ± 0%     118MB ± 0%   -6.28%  (p=1.000 n=1+1)
Compaction/type=normal,blocks=4,series=10000,samplesPerSeriesPerBlock=2001-4       163MB ± 0%     146MB ± 0%  -10.22%  (p=1.000 n=1+1)
Compaction/type=normal,blocks=4,series=10000,samplesPerSeriesPerBlock=5001-4       295MB ± 0%     251MB ± 0%  -14.75%  (p=1.000 n=1+1)
Compaction/type=vertical,blocks=4,series=10000,samplesPerSeriesPerBlock=101-4      154MB ± 0%     153MB ± 0%   -0.72%  (p=1.000 n=1+1)
Compaction/type=vertical,blocks=4,series=10000,samplesPerSeriesPerBlock=1001-4     465MB ± 0%     455MB ± 0%   -2.10%  (p=1.000 n=1+1)
Compaction/type=vertical,blocks=4,series=10000,samplesPerSeriesPerBlock=2001-4     755MB ± 0%     742MB ± 0%   -1.81%  (p=1.000 n=1+1)
Compaction/type=vertical,blocks=4,series=10000,samplesPerSeriesPerBlock=5001-4    1.73GB ± 0%    1.69GB ± 0%   -2.03%  (p=1.000 n=1+1)

name                                                                            old allocs/op  new allocs/op  delta
Compaction/type=normal,blocks=4,series=10000,samplesPerSeriesPerBlock=101-4        1.33M ± 0%     1.25M ± 0%   -6.01%  (p=1.000 n=1+1)
Compaction/type=normal,blocks=4,series=10000,samplesPerSeriesPerBlock=1001-4       2.29M ± 0%     1.57M ± 0%  -31.43%  (p=1.000 n=1+1)
Compaction/type=normal,blocks=4,series=10000,samplesPerSeriesPerBlock=2001-4       3.25M ± 0%     1.89M ± 0%  -41.83%  (p=1.000 n=1+1)
Compaction/type=normal,blocks=4,series=10000,samplesPerSeriesPerBlock=5001-4       6.25M ± 0%     2.89M ± 0%  -53.75%  (p=1.000 n=1+1)
Compaction/type=vertical,blocks=4,series=10000,samplesPerSeriesPerBlock=101-4      1.79M ± 0%     1.71M ± 0%   -4.47%  (p=1.000 n=1+1)
Compaction/type=vertical,blocks=4,series=10000,samplesPerSeriesPerBlock=1001-4     5.05M ± 0%     4.33M ± 0%  -14.25%  (p=1.000 n=1+1)
Compaction/type=vertical,blocks=4,series=10000,samplesPerSeriesPerBlock=2001-4     8.08M ± 0%     6.72M ± 0%  -16.83%  (p=1.000 n=1+1)
Compaction/type=vertical,blocks=4,series=10000,samplesPerSeriesPerBlock=5001-4     18.0M ± 0%     14.6M ± 0%  -18.69%  (p=1.000 n=1+1)```